Create Final with improvisation
The proper use of input validation loop and the forever while loop. The input validation loop played a major part in this lab. It helped the user only to enter valid numbers and also alert him if the input was incorrect with a proper escape sequence and changing the console color. The proper format was followed using the hierarchy chart and proper pseudocode. The conditional expression helped me to write concise lines of code which was helpful while reading the code. The proper way of capitalizing an alphabet without using the ASCII numbers was something new I learned from this lab. The logical operators were useful while differentiating between digits, special characters, and alphabets. It was helpful for capitalizing the alphabets as well, so they check in the switch case. The member functions or methods helped to clean up the input. The lessons learned were the proper use of the conditional operator and input validation loop. The switch case statements were useful while writing menu- driven programs.